commit 3:
In the "Artists" tab on the excel sheet (Spotify2023_artists df) The following updates were made: De-dupped artist and summed total streams by artist; added rank column to the data frame to rank artists by streams. This will allow rank to act as a card visual in Tableau. Reduced the rows to the top 100 artists for simplicity purposes.

commit 3.1: filled in null values in the 'key' column